Endow framework to control threads per workload
Remove from worker the ability to select randomly the workload.
It enables the framework to control / specify the number of threads
for each registered workload. It's better for debug purposes, particularly
for debugging the behavior with and without interaction of a given workload
alone or with other workload of its own kind.

It also improves worker output in debug mode putting all the results in one
single line, and fixes some small typos in the output...