torture/narrow: improve context switch test-case
Apparently there is yet another issue. I suspect it's precisely
when KVM reschedule happens and the HTM transaction is reclaimed/
recheckpointed. I introduced a print of a dot when such a condition
happens so at least we can have a clue when we see a corruption. If
KVM is the culprit then a dot must always be present (tho such a
condition is very hard to catch :-( ).