narrow/signal: Add simplified form for trap issue
This is a most simplified form for the illegal instruction caugh
if we do a `trap` in TM with MSR_FP and MSR_VEC both disabled. In
that case instead of taking the failure handler (rollbacking fine)
we take an illegal instruction pointing to `beq <failure_handler>`
which is not the expected behavior.